Quantize band texture to Rgba16Sint (Stage B of RGBA16I migration)
Switch band texture from Rgba32Uint (16 bytes/texel) to Rgba16Sint
(8 bytes/texel). Band headers, curve refs, and split coordinates all
fit in i16. Splits are quantized at 4 units/em matching curve data.

Both textures are now Rgba16Sint — total texture memory is halved
compared to the mixed Rgba32Float/Rgba32Uint baseline.

Coded by an LLM.

+++ tests/debug_bold_r.rs
@@ -81,7 +81,7 @@ fn dump_bold_r_geometry() {
             for ci in 0..count {
                 let ref_base = (offset + ci) * 4;
                 let curve_x = band_data.entries[ref_base];
-                if curve_x / 2 == target_curve {
+                if curve_x / 2 == target_curve as i16 {
                     h_bands.push(i);
                 }
             }
@@ -94,7 +94,7 @@ fn dump_bold_r_geometry() {
             for ci in 0..count {
                 let ref_base = (offset + ci) * 4;
                 let curve_x = band_data.entries[ref_base];
-                if curve_x / 2 == target_curve {
+                if curve_x / 2 == target_curve as i16 {
                     v_bands.push(i);
                 }
             }
